Move integration tests from functional/ to integration/
