+++ extensions/amp-bind/0.1/test/integration/test-bind-impl.js
@@ -15,12 +15,12 @@
  */
 
 import * as sinon from 'sinon';
-import {AmpEvents} from '../../../../src/amp-events';
-import {Bind} from '../bind-impl';
-import {BindEvents} from '../bind-events';
-import {chunkInstanceForTesting} from '../../../../src/chunk';
-import {toArray} from '../../../../src/types';
-import {user} from '../../../../src/log';
+import {AmpEvents} from '../../../../../src/amp-events';
+import {Bind} from '../../bind-impl';
+import {BindEvents} from '../../bind-events';
+import {chunkInstanceForTesting} from '../../../../../src/chunk';
+import {toArray} from '../../../../../src/types';
+import {user} from '../../../../../src/log';
 
 /**
  * @param {!Object} env
