+++ extensions/amp-viewer-integration/0.1/test/integration/test-amp-webview-viewer-integration.js
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-import {
-    WebviewViewerForTesting,
-} from './viewer-initiated-handshake-viewer-for-testing';
+import {AmpViewerIntegration} from '../../amp-viewer-integration';
+import {WebviewViewerForTesting} from '../webview-viewer-for-testing.js';
 
 
 describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
@@ -33,6 +32,7 @@ describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
       viewerEl = document.createElement('div');
       document.body.appendChild(viewerEl);
       viewer = new WebviewViewerForTesting(viewerEl, '1', ampDocUrl, true);
+      return viewer.waitForHandshakeResponse();
     });
 
     afterEach(() => {
@@ -40,15 +40,38 @@ describes.sandboxed('AmpWebviewViewerIntegration', {}, () => {
     });
 
     it('should confirm the handshake', () => {
-      return viewer.waitForHandshakeResponse();
+      console/*OK*/.log('sending handshake response');
+      return viewer.waitForDocumentLoaded();
     });
 
     it('should handle unload correctly', () => {
-      viewer.waitForHandshakeResponse().then(() => {
+      viewer.waitForDocumentLoaded().then(() => {
         const stub = sandbox.stub(viewer, 'handleUnload_');
         window.eventListeners.fire({type: 'unload'});
         expect(stub).to.be.calledOnce;
       });
+    });
+  });
+
+  describes.fakeWin('webview window init', {
+    amp: {
+      params: {
+        webview: '1',
+        origin: null,
+      },
+    },
+  }, env => {
+    let integr;
+
+    beforeEach(() => {
+      integr = new AmpViewerIntegration(env.win);
+    });
+
+    it('should set source and origin for webview', () => {
+      const stub = sandbox.stub(integr, 'webviewPreHandshakePromise_',
+          () => new Promise(() => {}));
+      integr.init();
+      expect(stub).to.be.calledWith(/* source */ null, /* origin */ '');
     });
   });
 });
